requirements.txt
requests>=2.31.0
pyyaml>=6.0

# Testing only
vcrpy>=6.0.0
pytest>=7.0.0
